Modified the creation menu items to be more compartmentalized, setup the gesture time controller in the creation of the object.
Worked a bit on planning the compilation of the colliders in the gesture script. Planned out some potential abilities of the gesture settings.
Rewrote some of the Gesture time controller script to be able to control the colliders within the gesture, rather than rely on the manager outside to do so. Will need additional work to function properly, and will need bug testing.